Move security to service layer
