+++ app/service/auth.py
@@ -6,8 +6,8 @@ from fastapi import Depends, HTTPException, status
 from jwt import PyJWTError
 
 from app.api.api_v1.schemas.user import JWTUser
-from app.core import security
 from app.repository.user import get_app_user_by_email
+from app.service import security
 
 CREDENTIALS_EXCEPTION = HTTPException(
     status_code=status.HTTP_401_UNAUTHORIZED,
